save to SD card and water level calculation.
Transitioned the "save to SD card" script into the water level attempt script. Also calculated water level by subtracting water pressure from barometric pressure. After a test with a water bottle with 1 cm markings on it, it's evident that the sensors are pretty sensitive, so even when the water pressure sensor is stationary the calculation says that pressure is changing (but it's not). The calculation is close to 1 cm/1 mbar. I'm also not sure if I updated the repository correctly.